fix no fish

+++ cohorts_proj/api/adapters/dar.py
@@ -162,8 +162,8 @@ def get_dataframe_nofish():
                 'PNFFQTUNA']
 
     for var in fishvars:
-        df_new[var] = df_new[var].astype(int)
-
+        df_new[var] = df_new[var].astype(float)
+    
     dar_logic = (df_new['PNFFQFR_FISH_KIDS'] == 0) & (df_new['PNFFQDK_FISH'] == 0) & (df_new['fish'] == 0)  & (df_new['PNFFQSHRIMP_CKD'] == 0)  & (df_new['PNFFQSHRIMP_CKD'] == 0) & (df_new['TOTALFISH_SERV'] == 0) & (df_new['PNFFQTUNA'] == 0)
 
     df_nofish = df_new[dar_logic]
